refactor: moved in new file server, client and dispatcher interfaces
