+++ dispatcher.go
@@ -1,0 +1,90 @@
+package osc
+
+import (
+	"fmt"
+	"strings"
+	"time"
+)
+
+// Dispatcher is an interface for an OSC message dispatcher. A dispatcher is
+// responsible for dispatching received OSC messages.
+type Dispatcher interface {
+	Dispatch(packet Packet)
+}
+
+// StandardDispatcher is a dispatcher for OSC packets. It handles the dispatching of
+// received OSC packets to Handlers for their given address.
+type StandardDispatcher struct {
+	handlers       map[string]Handler
+	defaultHandler Handler
+}
+
+// NewStandardDispatcher returns an StandardDispatcher.
+func NewStandardDispatcher() *StandardDispatcher {
+	return &StandardDispatcher{
+		handlers:       make(map[string]Handler),
+		defaultHandler: nil,
+	}
+}
+
+// AddMsgHandler adds a new message handler for the given OSC address.
+func (s *StandardDispatcher) AddMsgHandler(addr string, handler HandlerFunc) error {
+	if addr == "*" {
+		s.defaultHandler = handler
+		return nil
+	}
+
+	for _, chr := range "*?,[]{}# " {
+		if strings.Contains(addr, fmt.Sprintf("%c", chr)) {
+			return ERROR_OSC_INVALID_CHARACTER
+		}
+	}
+
+	if addressExists(addr, s.handlers) {
+		return ERROR_OSC_ADDRESS_EXISTS
+	}
+
+	s.handlers[addr] = handler
+
+	return nil
+}
+
+// Dispatch dispatches OSC packets. Implements the Dispatcher interface.
+func (s *StandardDispatcher) Dispatch(packet Packet) {
+	switch p := packet.(type) {
+	case *Message:
+		for addr, handler := range s.handlers {
+			if p.Match(addr) {
+				handler.HandleMessage(p)
+			}
+		}
+
+		if s.defaultHandler != nil {
+			s.defaultHandler.HandleMessage(p)
+		}
+
+	case *Bundle:
+		timer := time.NewTimer(p.Timetag.ExpiresIn())
+
+		go func() {
+			<-timer.C
+
+			for _, message := range p.Messages {
+				for address, handler := range s.handlers {
+					if message.Match(address) {
+						handler.HandleMessage(message)
+					}
+				}
+
+				if s.defaultHandler != nil {
+					s.defaultHandler.HandleMessage(message)
+				}
+			}
+
+			// Process all bundles
+			for _, b := range p.Bundles {
+				s.Dispatch(b)
+			}
+		}()
+	}
+}
