+++ osc.go
@@ -7,10 +7,8 @@ import (
 	"bytes"
 	"encoding/binary"
 	"fmt"
-	"net"
 	"regexp"
 	"strings"
-	"time"
 )
 
 const (
@@ -21,29 +19,6 @@ const (
 // Packet is the interface for Message and Bundle.
 type Packet interface {
 	MarshalBinary() (data []byte, err error)
-}
-
-// Client enables you to send OSC packets. It sends OSC messages and bundles to
-// the given IP address and port.
-type Client struct {
-	IP    string
-	Port  int
-	laddr *net.UDPAddr
-}
-
-// Server represents an OSC server. The server listens on Address and Port for
-// incoming OSC packets and bundles.
-type Server struct {
-	Addr        string
-	Dispatcher  Dispatcher
-	ReadTimeout time.Duration
-	close       func() error
-}
-
-// Dispatcher is an interface for an OSC message dispatcher. A dispatcher is
-// responsible for dispatching received OSC messages.
-type Dispatcher interface {
-	Dispatch(packet Packet)
 }
 
 // Handler is an interface for message handlers. Every handler implementation
@@ -65,214 +40,6 @@ func (f HandlerFunc) HandleMessage(msg *Message) {
 ////
 // StandardDispatcher
 ////
-
-// StandardDispatcher is a dispatcher for OSC packets. It handles the dispatching of
-// received OSC packets to Handlers for their given address.
-type StandardDispatcher struct {
-	handlers       map[string]Handler
-	defaultHandler Handler
-}
-
-// NewStandardDispatcher returns an StandardDispatcher.
-func NewStandardDispatcher() *StandardDispatcher {
-	return &StandardDispatcher{
-		handlers:       make(map[string]Handler),
-		defaultHandler: nil,
-	}
-}
-
-// AddMsgHandler adds a new message handler for the given OSC address.
-func (s *StandardDispatcher) AddMsgHandler(addr string, handler HandlerFunc) error {
-	if addr == "*" {
-		s.defaultHandler = handler
-		return nil
-	}
-
-	for _, chr := range "*?,[]{}# " {
-		if strings.Contains(addr, fmt.Sprintf("%c", chr)) {
-			return ERROR_OSC_INVALID_CHARACTER
-		}
-	}
-
-	if addressExists(addr, s.handlers) {
-		return ERROR_OSC_ADDRESS_EXISTS
-	}
-
-	s.handlers[addr] = handler
-
-	return nil
-}
-
-// Dispatch dispatches OSC packets. Implements the Dispatcher interface.
-func (s *StandardDispatcher) Dispatch(packet Packet) {
-	switch p := packet.(type) {
-	case *Message:
-		for addr, handler := range s.handlers {
-			if p.Match(addr) {
-				handler.HandleMessage(p)
-			}
-		}
-
-		if s.defaultHandler != nil {
-			s.defaultHandler.HandleMessage(p)
-		}
-
-	case *Bundle:
-		timer := time.NewTimer(p.Timetag.ExpiresIn())
-
-		go func() {
-			<-timer.C
-
-			for _, message := range p.Messages {
-				for address, handler := range s.handlers {
-					if message.Match(address) {
-						handler.HandleMessage(message)
-					}
-				}
-
-				if s.defaultHandler != nil {
-					s.defaultHandler.HandleMessage(message)
-				}
-			}
-
-			// Process all bundles
-			for _, b := range p.Bundles {
-				s.Dispatch(b)
-			}
-		}()
-	}
-}
-
-////
-// Client
-////
-
-// NewClient creates a new OSC client. The Client is used to send OSC
-// messages and OSC bundles over an UDP network connection. The `ip` argument
-// specifies the IP address and `port` defines the target port where the
-// messages and bundles will be send to.
-func NewClient(ip string, port int) *Client {
-	return &Client{
-		IP:    ip,
-		Port:  port,
-		laddr: nil,
-	}
-}
-
-// SetLocalAddr sets the local address.
-func (c *Client) SetLocalAddr(ip string, port int) error {
-	laddr, err := net.ResolveUDPAddr("udp", fmt.Sprintf("%s:%d", ip, port))
-	if err != nil {
-		return err
-	}
-
-	c.laddr = laddr
-
-	return nil
-}
-
-// Send sends an OSC Bundle or an OSC Message.
-func (c *Client) Send(packet Packet) error {
-	addr, err := net.ResolveUDPAddr("udp", fmt.Sprintf("%s:%d", c.IP, c.Port))
-	if err != nil {
-		return err
-	}
-
-	conn, err := net.DialUDP("udp", c.laddr, addr)
-	if err != nil {
-		return err
-	}
-
-	defer conn.Close()
-
-	data, err := packet.MarshalBinary()
-	if err != nil {
-		return err
-	}
-
-	_, err = conn.Write(data)
-
-	return err
-}
-
-////
-// Server
-////
-
-// ListenAndServe retrieves incoming OSC packets and dispatches the retrieved
-// OSC packets.
-func (s *Server) ListenAndServe() error {
-	defer s.Close()
-
-	if s.Dispatcher == nil {
-		s.Dispatcher = NewStandardDispatcher()
-	}
-
-	ln, err := net.ListenPacket("udp", s.Addr)
-	if err != nil {
-		return err
-	}
-
-	s.close = ln.Close
-
-	return s.Serve(ln)
-}
-
-// Serve retrieves incoming OSC packets from the given connection and dispatches
-// retrieved OSC packets. If something goes wrong an error is returned.
-func (s *Server) Serve(c net.PacketConn) error {
-	tempDelay := 25 + time.Millisecond
-
-	for {
-		msg, err := s.Read(c)
-		if err != nil {
-			ne, ok := err.(net.Error)
-
-			if ok && ne.Temporary() {
-				time.Sleep(tempDelay)
-				continue
-			}
-
-			return err
-		}
-
-		go s.Dispatcher.Dispatch(msg)
-	}
-}
-
-// Close forcibly closes a server's connection.
-//
-// This causes a "use of closed network connection" error the next time the
-// server attempts to read from the connection.
-func (s *Server) Close() error {
-	if s.close == nil {
-		return nil
-	}
-
-	return s.close()
-}
-
-// Read retrieves OSC packets.
-func (s *Server) Read(c net.PacketConn) (Packet, error) {
-	if s.ReadTimeout != 0 {
-		err := c.SetReadDeadline(time.Now().Add(s.ReadTimeout))
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	data := make([]byte, 65535)
-
-	n, _, err := c.ReadFrom(data)
-	if err != nil {
-		return nil, err
-	}
-
-	var start int
-	p, err := readPacket(bufio.NewReader(bytes.NewBuffer(data)), &start, n)
-
-	return p, err
-}
 
 // receivePacket receives an OSC packet from the given reader.
 func readPacket(reader *bufio.Reader, start *int, end int) (Packet, error) {
